investika/requirements.txt
Pillow==9.0.1
pytz==2022.1
rest-framework-simplejwt==0.0.2
zipp==1.0.0
authlib==1.0
django==4.2
python-dotenv==0.19
requests==2.27
python-decouple==3.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
black==23.7.0 
